# openapi.json
{"openapi":"3.0.1","info":{"title":"ScotWest Bank Enterprise Core API","description":"Production-grade core banking backend covering Customer Management, Account Services, Atomic Transfers, and Audit Ledger.","contact":{"name":"ScotWest Engineering Team","email":"dev@scotwest.co.uk"},"license":{"name":"Apache 2.0","url":"https://springdoc.org"},"version":"v1.0.0"},"servers":[{"url":"http://localhost:8085","description":"Generated server url"}],"paths":{"/api/v1/transfers":{"post":{"tags":["transfer-controller"],"operationId":"transferFunds","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResponse"}}}}}}},"/api/v1/customers":{"get":{"tags":["customer-controller"],"operationId":"getAllCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}}},"post":{"tags":["customer-controller"],"operationId":"registerCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/api/v1/accounts":{"post":{"tags":["account-controller"],"operationId":"openAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/api/v1/transfers/history/{accountNumber}":{"get":{"tags":["transfer-controller"],"operationId":"getHistory","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/v1/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getCustomerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/api/v1/accounts/{accountNumber}":{"get":{"tags":["account-controller"],"operationId":"getAccountByNumber","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/api/v1/accounts/customer/{customerId}":{"get":{"tags":["account-controller"],"operationId":"getAccountsByCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"}}}}}}}}},"components":{"schemas":{"TransferRequest":{"required":["amount","sourceAccountNumber","targetAccountNumber"],"type":"object","properties":{"sourceAccountNumber":{"type":"string"},"targetAccountNumber":{"type":"string"},"amount":{"minimum":0.01,"exclusiveMinimum":false,"type":"number"},"description":{"type":"string"}}},"TransferResponse":{"type":"object","properties":{"transactionReference":{"type":"string"},"sourceAccountNumber":{"type":"string"},"targetAccountNumber":{"type":"string"},"amount":{"type":"number"},"status":{"type":"string"},"description":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"CustomerRequest":{"required":["email","firstName","lastName","phoneNumber"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"}}},"Customer":{"required":["email","firstName","lastName"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"AccountRequest":{"required":["accountType","customerId","initialDeposit"],"type":"object","properties":{"customerId":{"type":"integer","format":"int64"},"accountType":{"type":"string"},"initialDeposit":{"minimum":0.0,"exclusiveMinimum":false,"type":"number"}}},"AccountResponse":{"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"accountType":{"type":"string"},"balance":{"type":"number"},"customerId":{"type":"integer","format":"int64"},"customerName":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"Transaction":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"transactionReference":{"type":"string"},"sourceAccountNumber":{"type":"string"},"targetAccountNumber":{"type":"string"},"amount":{"type":"number"},"status":{"type":"string"},"description":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}}}}}