# src/main/java/com/scotwest/banking/dto/CustomerRequest.java
package com.scotwest.banking.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class CustomerRequest {

    @NotBlank(message = "First name cannot be blank")
    private String firstName;

    @NotBlank(message = "Last name cannot be blank")
    private String lastName;

    @NotBlank(message = "Email cannot be blank")
    @Email(message = "Provide a valid email address")
    private String email;

    @NotBlank(message = "Phone number cannot be blank")
    private String phoneNumber;
}